Enable resources to have a "docker_client" configured
This allows containers to be provisioned on different remote hosts
