+++ internal/provider/resource_docker_buildx_builder.go
@@ -31,6 +31,7 @@ func resourceDockerBuildxBuilder() *schema.Resource {
 		DeleteContext: resourceDockerBuildxBuilderDelete,
 		Description:   "Manages a Docker Buildx builder instance. This resource allows you to create a  buildx builder with various configurations such as driver, nodes, and platform settings. Please see https://github.com/docker/buildx/blob/master/docs/reference/buildx_create.md for more documentation",
 		Schema: map[string]*schema.Schema{
+			"docker_client": dockerSchema,
 			"name": {
 				Type:        schema.TypeString,
 				Default:     "",
@@ -429,7 +430,11 @@ func resourceDockerBuildxBuilderCreate(ctx context.Context, d *schema.ResourceDa
 		driverOptions = processDriverOptions(remote)
 	}
 
-	client := meta.(*ProviderConfig).DockerClient
+	client, err := NewDockerClient(ctx, d)
+
+	if err != nil {
+		client = meta.(*ProviderConfig).DockerClient
+	}
 
 	t, error := command.NewDockerCli()
 	if error != nil {
@@ -437,7 +442,7 @@ func resourceDockerBuildxBuilderCreate(ctx context.Context, d *schema.ResourceDa
 	}
 
 	log.Printf("[DEBUG] Docker CLI initialized %#v, %#v", client, client.DaemonHost())
-	err := t.Initialize(&flags.ClientOptions{Hosts: []string{client.DaemonHost()}})
+	err = t.Initialize(&flags.ClientOptions{Hosts: []string{client.DaemonHost()}})
 
 	if err != nil {
 		return diag.FromErr(fmt.Errorf("failed to initialize Docker CLI: %w", err))
@@ -520,13 +525,17 @@ func processDriverOptions(driverOptionsMap map[string]interface{}) []string {
 // resourceDockerBuildxBuilderRead handles reading the state of a Buildx builder
 // corresponding file in buildx repo: https://github.com/docker/buildx/blob/master/commands/inspect.go
 func resourceDockerBuildxBuilderRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	client := meta.(*ProviderConfig).DockerClient
+	client, err := NewDockerClient(ctx, d)
+
+	if err != nil {
+		client = meta.(*ProviderConfig).DockerClient
+	}
 
 	dockerCli, error := command.NewDockerCli()
 	if error != nil {
 		return diag.FromErr(fmt.Errorf("failed to create Docker CLI: %w", error))
 	}
-	err := dockerCli.Initialize(&flags.ClientOptions{Hosts: []string{client.DaemonHost()}})
+	err = dockerCli.Initialize(&flags.ClientOptions{Hosts: []string{client.DaemonHost()}})
 	if err != nil {
 		return diag.FromErr(fmt.Errorf("failed to initialize Docker CLI: %w", err))
 	}
@@ -550,7 +559,11 @@ func resourceDockerBuildxBuilderDelete(ctx context.Context, d *schema.ResourceDa
 	name := d.Id()
 	log.Printf("[DEBUG] Deleting Buildx builder: %s", name)
 
-	client := meta.(*ProviderConfig).DockerClient
+	client, err := NewDockerClient(ctx, d)
+
+	if err != nil {
+		client = meta.(*ProviderConfig).DockerClient
+	}
 
 	dockerCli, err := command.NewDockerCli()
 	if err != nil {
