+++ internal/provider/resource_docker_registry_image.go
@@ -14,6 +14,7 @@ func resourceDockerRegistryImage() *schema.Resource {
 		UpdateContext: resourceDockerRegistryImageUpdate,
 
 		Schema: map[string]*schema.Schema{
+			"docker_client": dockerSchema,
 			"name": {
 				Type:        schema.TypeString,
 				Description: "The name of the Docker image.",
